requirements.txt
fastapi
uvicorn
gunicorn
redis
rq
SQLAlchemy
psycopg2-binary
python-dotenv
cloudinary
google-api-python-client
google-auth
tenacity
requests
pydantic
